moved calculate_multiple_hisim_requests into separate module
